homeworks/projects/Movie_rental/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
pydantic==1.10.2
